+++ feature/src/main/java/com/eoyeongbooyeong/search/component/PlaceInfoListItem.kt
@@ -45,7 +45,7 @@ import com.eoyeongbooyeong.feature.R
 internal fun PlaceInfoListItem(
     placeName: String,
     address: String,
-    star: Float,
+    star: Double,
     reviewCount: Int,
     likedCount: Int,
     movieNameList: List<String>?,
@@ -193,7 +193,7 @@ internal fun PlaceInfoListItem(
 @Composable
 fun PlaceReviewAndLikedCount(
     modifier: Modifier = Modifier,
-    star: Float = 0f,
+    star: Double = 0.0,
     reviewCount: Int = 0,
     likedCount: Int = 0,
 ) {
@@ -245,7 +245,7 @@ fun PlaceInfoResultListPreview() {
                 placeName = "PlaceEntity Name",
                 placeImageUrl = "https://placeimg.com/100/100/any",
                 address = "Address",
-                star = 4.5f,
+                star = 4.5,
                 reviewCount = 100,
                 likedCount = 100,
                 isBookmarked = false,
